pyproject.toml
[tool.poetry]
name = "motor_insurance_risk"
version = "0.1.0"
description = ""
authors = ["Son Nguyen <vanson.nguyen@covergo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
psycopg2-binary = "^2.9.10"
openai = "^1.52.1"
pgvector = "^0.3.5"
numpy = "^2.1.2"
pandas = "^2.2.3"
tiktoken = "^0.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
